# [WIP] Refactor global namespace into encapsulated application class