--- a/samples/app.cr
+++ b/samples/app.cr
@@ -1,0 +1,9 @@
+require "kemal/base"
+
+class MyApp < Kemal::Application
+  get "/" do
+    "Hello Kemal!"
+  end
+end
+
+MyApp.run