--- a/src/kemal/base.cr
+++ b/src/kemal/base.cr
@@ -1,4 +1,5 @@
 require "./helpers/*"
+require "./base/*"
 
 # Kemal Base
 # The DSL currently consists of
@@ -10,19 +11,18 @@ class Kemal::Base
   include FileHelpers
   include Templates
   include Macros
+  include Base::DSL
+  include Base::Builder
 
-  HTTP_METHODS   = %w(get post put patch delete options)
-  FILTER_METHODS = %w(get post put patch delete options all)
-
+  # :nodoc:
   getter route_handler = Kemal::RouteHandler.new
+  # :nodoc:
   getter filter_handler = Kemal::FilterHandler.new
+  # :nodoc:
   getter websocket_handler = Kemal::WebSocketHandler.new
 
   getter handlers = [] of HTTP::Handler
-  getter custom_handlers = [] of Tuple(Nil | Int32, HTTP::Handler)
-  getter filter_handlers = [] of HTTP::Handler
   getter error_handlers = {} of Int32 => HTTP::Server::Context, Exception -> String
-  @handler_position = 0
 
   getter config : Config
 
@@ -31,133 +31,11 @@ class Kemal::Base
   property? running = false
 
   def initialize(@config = Config.base)
-    @logger = if @config.logging?
-                Kemal::LogHandler.new
-              else
-                Kemal::NullLogHandler.new
-              end
-    add_filter_handler(filter_handler)
-  end
+    @filter_handler = FilterHandler.new(self)
+    @route_handler = RouteHandler.new(self)
+    @websocket_handler = WebSocketHandler.new(self)
 
-  {% for method in HTTP_METHODS %}
-    def {{method.id}}(path, &block : HTTP::Server::Context -> _)
-      raise Kemal::Exceptions::InvalidPathStartException.new({{method}}, path) unless Kemal::Utils.path_starts_with_slash?(path)
-      route_handler.add_route({{method}}.upcase, path, &block)
-    end
-  {% end %}
-
-  def ws(path, &block : HTTP::WebSocket, HTTP::Server::Context -> Void)
-    raise Kemal::Exceptions::InvalidPathStartException.new("ws", path) unless Kemal::Utils.path_starts_with_slash?(path)
-    websocket_handler.add_route path, &block
-  end
-
-  def error(status_code, &block : HTTP::Server::Context, Exception -> _)
-    add_error_handler status_code, &block
-  end
-
-  # All the helper methods available are:
-  #  - before_all, before_get, before_post, before_put, before_patch, before_delete, before_options
-  #  - after_all, after_get, after_post, after_put, after_patch, after_delete, after_options
-  {% for type in ["before", "after"] %}
-    {% for method in FILTER_METHODS %}
-      def {{type.id}}_{{method.id}}(path = "*", &block : HTTP::Server::Context -> _)
-        filter_handler.{{type.id}}({{method}}.upcase, path, &block)
-      end
-    {% end %}
-  {% end %}
-
-  def clear
-    @router_included = false
-    @handler_position = 0
-    @default_handlers_setup = false
-
-    handlers.clear
-    custom_handlers.clear
-    filter_handlers.clear
-    error_handlers.clear
-
-    route_handler.clear
-    websocket_handler.clear
-  end
-
-  def handlers=(handlers : Array(HTTP::Handler))
-    clear
-    @handlers.replace(handlers)
-  end
-
-  def add_handler(handler : HTTP::Handler)
-    @custom_handlers << {nil, handler}
-  end
-
-  def add_handler(handler : HTTP::Handler, position : Int32)
-    @custom_handlers << {position, handler}
-  end
-
-  def add_filter_handler(handler : HTTP::Handler)
-    @filter_handlers << handler
-  end
-
-  def add_error_handler(status_code, &handler : HTTP::Server::Context, Exception -> _)
-    @error_handlers[status_code] = ->(context : HTTP::Server::Context, error : Exception) { handler.call(context, error).to_s }
-  end
-
-  def setup
-    unless @default_handlers_setup && @router_included
-      setup_init_handler
-      setup_log_handler
-      setup_error_handler
-      setup_static_file_handler
-      setup_custom_handlers
-      setup_filter_handlers
-      @default_handlers_setup = true
-      @router_included = true
-      handlers.insert(handlers.size, websocket_handler)
-      handlers.insert(handlers.size, route_handler)
-    end
-  end
-
-  private def setup_init_handler
-    @handlers.insert(@handler_position, Kemal::InitHandler.new(self))
-    @handler_position += 1
-  end
-
-  private def setup_log_handler
-    @handlers.insert(@handler_position, logger)
-    @handler_position += 1
-  end
-
-  private def setup_error_handler
-    if @config.always_rescue?
-      @error_handler ||= Kemal::ExceptionHandler.new
-      @handlers.insert(@handler_position, @error_handler.not_nil!)
-      @handler_position += 1
-    end
-  end
-
-  private def setup_static_file_handler
-    if @config.serve_static.is_a?(Hash)
-      @handlers.insert(@handler_position, Kemal::StaticFileHandler.new(@config.public_folder))
-      @handler_position += 1
-    end
-  end
-
-  private def setup_custom_handlers
-    @custom_handlers.each do |ch|
-      position = ch[0]
-      if !position
-        @handlers.insert(@handler_position, ch[1])
-        @handler_position += 1
-      else
-        @handlers.insert(position, ch[1])
-        @handler_position += 1
-      end
-    end
-  end
-
-  private def setup_filter_handlers
-    @filter_handlers.each do |h|
-      @handlers.insert(@handler_position, h)
-    end
+    initialize_defaults
   end
 
   # Overload of self.run with the default startup logging
@@ -178,6 +56,11 @@ class Kemal::Base
     start_server(port) do
       yield self
     end
+  end
+
+  # DEPRECATED: This method should be replaced with `#running?`
+  def running
+    running?
   end
 
   private def prepare_for_server_start