--- a/src/kemal/base.cr
+++ b/src/kemal/base.cr
@@ -99,4 +99,8 @@ class Kemal::Base
       raise "Kemal is already stopped."
     end
   end
+
+  def log(message)
+    logger.write "#{message}\n"
+  end
 end