--- a/src/kemal/base.cr
+++ b/src/kemal/base.cr
@@ -15,18 +15,19 @@ class Kemal::Base
   include Base::Builder
 
   # :nodoc:
-  getter route_handler = Kemal::RouteHandler.new
+  # TODO: These ivars are initialized in the constructor, but their values depend on `self`.
+  getter! route_handler : RouteHandler?
   # :nodoc:
-  getter filter_handler = Kemal::FilterHandler.new
+  getter! filter_handler : FilterHandler?
   # :nodoc:
-  getter websocket_handler = Kemal::WebSocketHandler.new
+  getter! websocket_handler : WebSocketHandler?
 
   getter handlers = [] of HTTP::Handler
   getter error_handlers = {} of Int32 => HTTP::Server::Context, Exception -> String
 
   getter config : Config
 
-  property! logger : Kemal::BaseLogHandler
+  property! logger : BaseLogHandler
   property! server : HTTP::Server
   property? running = false