--- a/src/kemal/base/builder.cr
+++ b/src/kemal/base/builder.cr
@@ -71,7 +71,7 @@ class Kemal::Base
 
     private def setup_error_handler
       if @config.always_rescue?
-        error_handler = @error_handler ||= Kemal::ExceptionHandler.new
+        error_handler = @error_handler ||= Kemal::ExceptionHandler.new(self)
         @handlers.insert(@handler_position, error_handler)
         @handler_position += 1
       end
@@ -79,7 +79,7 @@ class Kemal::Base
 
     private def setup_static_file_handler
       if @config.serve_static.is_a?(Hash)
-        @handlers.insert(@handler_position, Kemal::StaticFileHandler.new(@config.public_folder))
+        @handlers.insert(@handler_position, Kemal::StaticFileHandler.new(@config))
         @handler_position += 1
       end
     end