--- a/src/kemal/cli.cr
+++ b/src/kemal/cli.cr
@@ -3,11 +3,10 @@ require "option_parser"
 module Kemal
   # Handles all the initialization from the command line.
   class CLI
-    def initialize
+    def initialize(@config : Config = Kemal.config)
       @ssl_enabled = false
       @key_file = ""
       @cert_file = ""
-      @config = Kemal.config
       read_env
       parse
       configure_ssl