Instead of relying on global state everywhere (like `Kemal::Config::INSTANCE` etc.), this PR refactors all global namespace DSL features into encapsulated types to make them reusable. This is based on my suggestion in comment [#372 (comment)](372#issuecomment-314086157).

This enables users do opt-out of global namespace pollution and allows to use multiple Kemal applications in the same program.

So far I have rebuilt the entire current API - with `Kemal.config` (and new `Kemal.application`) pointing to class variables.