@straight-shoota 2017-07-17
Changes are mostly internal, so real-world use-cases should not be affected unless they used any of the global state constants directly. Most specs did not even require any changes or just minor adjustments because auf assumptions on underlying behavior.

The only breaking change is probably the refactoring of `Kemal::Config` into `Kemal::Config` for configuration values and `Kemal::Base` for application state - which would probably become necessary at some point anyway without this PR.

The next step would be move the helpers from global to application namespace and replace them with wrappers pointing to `Kemal.application` in top level namespace. This would make the top level namespace DSL completely optional.