@sdogruyol 2017-07-17
I can't find a `Kemal::Application` class here. Instead `Kemal::Base` is being used and it's really long.