@straight-shoota 2017-07-17
@sdogruyol Yes, I have just pushed the commit where I extracted this class from `Kemal::Base`. They are mostly the same, `Kemal::Application` just has some more opinionated default behavior as is appropriate for `Kemal.application`. On the other hand `Kemal::Base` is more vanilla and makes no assumptions on expected behavior. This is modeled after `Sinatra::Base` and `Sinatra::Application`.