@sdogruyol 2017-07-17
Meanwhile this won't work

```crystal
require "./kemal"

class MyApp < Kemal::Application
  get "/" do
    "Hello world!"
  end
end

MyApp.run
```