@straight-shoota 2017-07-17
I haven't gotten there yet. I was primarily focused on keeping compatibility with the existing API.