@straight-shoota 2017-07-17
There are a few road blocks on the way, because many components are directly coupled to global state. But I'm getting there but the loosend couplings will surely need some cleanup later to make it look nice.