@straight-shoota 2017-07-20
So I think I have finally figured out a way to use class-level DSL with instance scoped blocks: The blocks are converted to method definitions and the setup later adds blocks with calls to those custom methods to the handler list. 
Maybe it's a bit hacky but user experience is great 🥇
@sdogruyol please take a look, your example should work now. I have also added an example with two apps listening simultaneously on two different ports.