@sdogruyol 2017-07-20
Hey @straight-shoota, thanks for this PR 👍 I need to go over and digest this. However this week i don't have much time, please be patient for a reply