@hak8or 2017-07-20
Just wanted to pop in and say this would be amazing in terms allowing much better organization of code.