@sdogruyol 2017-07-28
@straight-shoota I'll start reviewing this on the weekend. Could you please fix the conflicts 🙏