@sdogruyol 2017-07-28
I've started reviewing and just tried doing some custom config for an app like below

```
class MyApp < Kemal::Application
  def initialize(config)
    config.serve_static = false
    config.logging = false
  end

  get "/" do |variable|
    "App 1"
  end
end

app = MyApp.new(Kemal::Config.new)
app.run
```

This causes a segfault

```
Invalid memory access (signal 11) at address 0x40
[0x109ab711b] *CallStack::print_backtrace:Int32 +107
[0x109a9ca4c] __crystal_sigfault_handler +60
```