@straight-shoota 2017-07-28
You need to call super, otherwise `@config` will not be set. 

I'd expect the compiler to complain about this... Might be related to [crystal-lang/lang#4151](https://github.com/crystal-lang/crystal/issues/4151).

An idiomatic way to do this would be to overwrite `.new` instead of `#initialize`.