@sdogruyol 2017-07-28
```crystal
class MyApp < Kemal::Application
  def initialize(config)
    super
    config.serve_static = false
    config.logging = false
  end

  get "/" do |variable|
    "App 1"
  end
end

app = MyApp.new(Kemal::Config.base)
app.run
```

This crashes with a nil assertion error

```
Nil assertion failed (Exception)
0x101c77c15: *CallStack::unwind:Array(Pointer(Void)) at ??
0x101c77bb1: *CallStack#initialize:Array(Pointer(Void)) at ??
0x101c77b88: *CallStack::new:CallStack at ??
0x101c5c985: *raise<Exception>:NoReturn at ??
0x101c5c931: *raise<String>:NoReturn at ??
0x101cb3173: *Nil#not_nil!:NoReturn at ??
0x101d7824d: *MyApp@Kemal::Base#logger:Kemal::BaseLogHandler+ at ??
0x101d78171: *MyApp@Kemal::Base#log<String>:(File::PReader | Flate::Reader | Flate::Writer | Gzip::Reader | Gzip::Writer | HTTP::ChunkedContent | HTTP::Server::Response | HTTP::Server::Response::Output | HTTP::UnknownLengthContent | HTTP::WebSocket::Protocol::StreamIO | IO::ARGF | IO::Delimited | IO::FileDescriptor+ | IO::Hexdump | IO::Memory | IO::MultiWriter | IO::Sized+ | OpenSSL::DigestIO | OpenSSL::SSL::Socket+ | String::Builder | Zlib::Reader | Zlib::Writer | Nil) at ??
0x101c743c7: ~procProc(Nil)@src/kemal/application.cr:9 at ??
0x101d78312: *MyApp@Kemal::Base#run<Nil, &Proc(Nil)>:Nil at ??
0x101d78152: *MyApp@Kemal::Application#run<N0x101d78152: *MyApp@Kemal::Application#run<Nil>:Nil at ??
Unhandled exception in spawn:
End of file reached (IO::EOFError)
0x101c77c15: *CallStack::unwind:Array(Pointer(Void)) at ??
0x101c77bb1: *CallStack#initialize:Array(Pointer(Void)) at ??
0x101c77b88: *CallStack::new:CallStack at ??
0x101c6ebe5: *raise<IO::EOFError>:NoReturn at ??
0x101cb72d5: *IO::FileDescriptor+@IO#read_fully<Slice(UInt8)>:Int32 at ??
0x101cd4d7b: *IO::ByteFormat::LittleEndian::decode<Int32:Class, IO::FileDescriptor+>:Int32 at ??
0x101ca3f92: *Int32@Int::from_io<IO::FileDescriptor+, IO::ByteFormat::LittleEndian:Module>:Int32 at ??
0x101cb9eb2: *IO::FileDescriptor+@IO#read_bytes<Int32:Class>:Int32 at ??
0x101d03525: *Event::SignalHandler#run:NoReturn at ??
0x101c7010c: ~procProc(Nil)@/usr/local/Cellar/crystal-lang/0.23.1/src/event/signal_handler.cr:93 at ??
0x101c90214: *Fiber#run:(IO::FileDescriptor | Nil) at ??
0x101c70819: ~proc2Proc(Fiber, (IO::FileDescriptor | Nil))@/usr/local/Cellar/crystal-lang/0.23.1/src/fiber.cr:29 at ??
riptor | Nil))0x101d78095: *MyApp@Kemal::Application#run:Nil at ??
0x101c5c421: __crystal_main at ??
0x101c709d8: main at ??
```