@straight-shoota 2017-07-28
Strange, your example works fine on my setup. What are your requires?

You have still a mistake though, you should pass `config` to `super`, otherwise your custom `config` object will not be stored to `@custom`. And you don't need to disable `serve_static` and `logging`: `Config.base` has them disabled by default.