@sdogruyol 2017-07-28
You're right 👍  This needs a good documentation