@sdogruyol 2017-07-28
Oops, still get a nil assertion error with this

```crystal
class MyApp < Kemal::Application
  def initialize(config)
    super(config)
  end

  get "/" do |variable|
    "App 1"
  end
end
```