@straight-shoota 2017-07-28
Unfortunately I still can't reproduce this error. Are your requiring `kemal` or `kemal/base` or something else?