@sdogruyol 2017-07-28
Actually, this was my bad. It's working correctly @straight-shoota