@sdogruyol 2017-07-28
From my first review, I like this 👍  Good job @straight-shoota. However the docs will need to be updated if this gets merged 😄