@straight-shoota 2017-07-28
Thank you 🎉 The docs should actually not be wrong because the preexisting API does not really change, but the alternative way should be added, of course.