@sdogruyol 2017-07-29
From my initial benchmarks this PR introduces ~%20 performance penalty to Kemal.

Master

```
➜  ~ wrk -t 2 -c 100 -d 40 http://localhost:3001
Running 40s test @ http://localhost:3001
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.62ms  396.18us  12.00ms   77.04%
    Req/Sec    19.10k   587.69    21.11k    76.00%
  1520672 requests in 40.01s, 163.88MB read
Requests/sec:  38010.31
Transfer/sec:      4.10MB
```

This branch

```
➜  ~ wrk -t 2 -c 100 -d 40 http://localhost:3000
Running 40s test @ http://localhost:3000
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.06ms  488.18us  11.05ms   83.20%
    Req/Sec    24.30k     1.45k   30.20k    81.50%
  1934758 requests in 40.00s, 208.50MB read
Requests/sec:  48364.50
Transfer/sec:      5.21MB
```