@straight-shoota 2017-10-18
I've finally spun up a vps to get better results than on my local machine and tested with this app to see if accessing the params parser makes a difference or not.

```crystal
require "kemal"

logging false
serve_static false

get "/" do |context|
  context.params.to_s
end

Kemal.run
```

I've compiled it against master (0b9628df) and jm-application(e91f08f6) and run tests with `wrk -c 100 -d 10 http://localhost:3000/`:

master:
```
Running 10s test @ http://localhost:3000/
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   150.73ms  224.05ms   1.10s    83.80%
    Req/Sec    17.99k    11.25k   30.13k    66.94%
  286360 requests in 10.05s, 38.21MB read
Requests/sec:  28499.19
Transfer/sec:      3.80MB
```
PR:
```
Running 10s test @ http://localhost:3000/
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   175.42ms  286.35ms   1.60s    86.88%
    Req/Sec    16.35k    10.00k   28.92k    58.82%
  251665 requests in 10.03s, 33.60MB read
Requests/sec:  25095.87
Transfer/sec:      3.35MB
```

So essentially, this seems to disprove my theory. It doesn't seem to matter much if `app` is assigned to each context or not.